Requests for service check log created

+++ backend/r2d2/src/main/java/com/unitybars/r2d2/dao/CheckLogDao.java
@@ -11,7 +11,7 @@ import java.util.List;
 public interface CheckLogDao {
     long insertCheckLog(CheckLog checkLog);
 
-    CheckLog getCheckLogById(int id);
+    CheckLog getCheckLogById(long id);
 
     List<CheckLog> getAllCheckLogs();
 }
